Data.Fits.Array - multidimensional array parsing with massiv
array2d

useful vector functions

working basic

up to 5d list parse

parsing massiv

more efficient resize without lists

cleanup

tests

cleanup

converted to lazy bytestrings

refactor

merge

docs
